Reorganize Config classes.
