+++ src/Robo.php
@@ -4,7 +4,8 @@ namespace Robo;
 use League\Container\Container;
 use League\Container\ContainerInterface;
 use Robo\Common\ProcessExecutor;
-use Robo\Common\ConfigLoader;
+use Robo\Config\Config;
+use Robo\Config\YamlConfigLoader;
 use Symfony\Component\Console\Input\StringInput;
 use Symfony\Component\Console\Application as SymfonyApplication;
 use Symfony\Component\Process\Process;
@@ -103,7 +104,7 @@ class Robo
      */
     public static function loadConfiguration($config, $paths)
     {
-        $loader = new ConfigLoader();
+        $loader = new YamlConfigLoader();
         foreach ($paths as $path) {
             $loader->load($path);
         }
@@ -118,7 +119,7 @@ class Robo
      * @param null|\Symfony\Component\Console\Input\InputInterface $input
      * @param null|\Symfony\Component\Console\Output\OutputInterface $output
      * @param null|\Robo\Application $app
-     * @param null|\Robo\Config $config
+     * @param null|\Robo\Config\Config $config
      *
      * @return \League\Container\Container|\League\Container\ContainerInterface
      */
@@ -164,7 +165,7 @@ class Robo
      *
      * @param \League\Container\ContainerInterface $container
      * @param \Symfony\Component\Console\Application $app
-     * @param \Robo\Config $config
+     * @param \Robo\Config\Config $config
      * @param null|\Symfony\Component\Console\Input\InputInterface $input
      * @param null|\Symfony\Component\Console\Output\OutputInterface $output
      */
@@ -332,7 +333,7 @@ class Robo
     }
 
     /**
-     * @return \Robo\Config
+     * @return \Robo\Config\Config
      */
     public static function config()
     {
